+++ packages/tessera/lib/src/l10n/l10n_cs.dart
@@ -153,6 +153,9 @@ final class TesseraStringsCs extends TesseraStrings {
   String get sortDescending => 'Seřadit sestupně';
 
   @override
+  String get inheritSort => 'Stejné pořadí jako úroveň výše';
+
+  @override
   String get expandAll => 'Rozbalit vše';
 
   @override