+++ packages/tessera/test/l10n_test.dart
@@ -81,6 +81,7 @@ void main() {
         l.largeExpansionTitle,
         l.sortAscending,
         l.sortDescending,
+        l.inheritSort,
         l.expandAll,
         l.collapseAll,
         l.countLabel,